+++ ami_api/__init__.py
@@ -1,27 +1,24 @@
 import os
-
-from flask import Flask
+import threading
+from .libs.aws import get_all_aws_ami
+from time import sleep
+from flask import Flask, g
 from logging.config import dictConfig
 
-dictConfig({
-    'version': 1,
-    'formatters': {'default': {
-        'format': '[%(asctime)s] %(levelname)s in %(module)s: %(message)s',
-    }},
-    'handlers': {'wsgi': {
-        'class': 'logging.StreamHandler',
-        'stream': 'ext://flask.logging.wsgi_errors_stream',
-        'formatter': 'default'
-    }},
-    'root': {
-        'level': 'INFO',
-        'handlers': ['wsgi']
-    }
-})
-
+delay_time = 180 # 3 minutes
+    
 def create_app(test_config=None):
     # create and configure the app
     app = Flask(__name__, instance_relative_config=True)
+
+    # Start the threading to retrieve ami images
+    with app.app_context():
+        def aws_ami():
+            while True:
+                get_all_aws_ami(app.logger)
+                sleep(delay_time)
+        aws_cron = threading.Thread(target=aws_ami)
+        aws_cron.start()
 
     if test_config is None:
         # load the instance config, if it exists, when not testing